backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic[email]==2.4.2
python-multipart==0.0.6
email-validator==2.1.0